pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "Телеграм-бот \"Помодоро\""
authors = ["Николай Марусов <nmarusov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncio = "^3.4.3"
telethon = "^1.36.0"
load-dotenv = "^0.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
